Implementing distinct servers TCP and UDP and adding your inicialization with threads in main class
