+++ app/LaunchServer.cs
@@ -6,23 +6,12 @@ class LaunchServer
 {
     public static void Main(String[] args)
     {
-        var serverApp = new ServerConfiguration();
-        serverApp.Start();
-        
-        /*var builderPacket = new WritePacket();
-        builderPacket.Write(1);
-        builderPacket.Write(100);
-        builderPacket.Write("GUGA");
-        builderPacket.Write(float.Pi);
-        var packet = builderPacket.BuildPacket();
-        Console.WriteLine("Length of packet > " + packet.Length);
-    
-        var reader = new ReadPacket(packet);
-        Console.WriteLine(reader.ReadInt());
-        Console.WriteLine(reader.ReadInt());
-        Console.WriteLine(reader.ReadString(4));
-        Console.WriteLine(reader.ReadFloat());*/
+        var tcpServer = new TcpServerConfiguration();
+        tcpServer.Start();
 
+        var udpServer = new UdpServerConfiguration();
+        udpServer.Start();
+        
     }
 
 }
